Уборка тестового проекта
- Классы разнесены по файлам
- Выполнено форматирование
