+++ Money.Api.Tests/ApiClient/ApiUser.cs
@@ -1,0 +1,5 @@
+﻿namespace Money.Api.Tests.ApiClient;
+
+public class ApiUser
+{
+}
