+++ Money.Api.Tests/AuthData.cs
@@ -1,0 +1,11 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Money.Api.Tests;
+
+public record AuthData(
+    [property: JsonPropertyName("access_token")]
+    string AccessToken,
+    [property: JsonPropertyName("token_type")]
+    string TokenType,
+    [property: JsonPropertyName("expires_in")]
+    int ExpiresIn);
