+++ Money.Api.Tests/CategoryTests.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.AspNetCore.Mvc.Testing;
+﻿using Money.Api.Tests.ApiClient;
 
 namespace Money.Api.Tests;
 
@@ -7,11 +7,11 @@ public class CategoryTests
     [Test]
     public async Task GetModifiedTest()
     {
-        var apiClient = new CategoryClient(Integration.GetHttpClient(), Console.WriteLine);
-        var result = (await apiClient.Get(1)).IsSuccess().Content;
+        CategoryClient apiClient = new(Integration.GetHttpClient(), Console.WriteLine);
+        CategoryClient.GetCategoriesModel? result = (await apiClient.Get(1)).IsSuccess().Content;
 
         Assert.That(result, Is.Not.Null);
         Assert.That(result.Categories, Is.Not.Null);
-        Assert.That(result.Categories.Length, Is.GreaterThanOrEqualTo(1));
+        Assert.That(result.Categories, Is.Not.Empty);
     }
 }
