+++ Money.Api.Tests/CustomWebApplicationFactory.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Money.Api.Tests;
+
+public class CustomWebApplicationFactory<TProgram> : WebApplicationFactory<TProgram> where TProgram : class
+{
+    protected override void ConfigureWebHost(IWebHostBuilder builder)
+    {
+        string env = "Development";
+
+        IConfigurationRoot configRoot = new ConfigurationBuilder()
+            .SetBasePath(Directory.GetCurrentDirectory())
+            .AddJsonFile($"appsettings.{env}.json")
+            .Build();
+
+        builder.ConfigureServices(services =>
+        {
+            services.AddSingleton(configRoot);
+        });
+
+        builder.UseConfiguration(configRoot);
+        builder.UseContentRoot(Directory.GetCurrentDirectory());
+        builder.UseEnvironment(env);
+    }
+}
