+++ Money.Api.Tests/Integration.cs
@@ -1,60 +1,24 @@
 using System.Collections.Concurrent;
 using System.Net.Http.Json;
-using System.Text.Json.Serialization;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace Money.Api.Tests;
-
-public class CustomWebApplicationFactory<TProgram> : WebApplicationFactory<TProgram> where TProgram : class
-{
-    protected override void ConfigureWebHost(IWebHostBuilder builder)
-    {
-        string env = "Development";
-
-        IConfigurationRoot configRoot = new ConfigurationBuilder()
-            .SetBasePath(Directory.GetCurrentDirectory())
-            .AddJsonFile($"appsettings.{env}.json")
-            .Build();
-
-       builder.ConfigureServices(services =>
-       {
-           services.AddSingleton(configRoot);
-       });
-
-        builder.UseConfiguration(configRoot);
-        builder.UseContentRoot(Directory.GetCurrentDirectory());
-        builder.UseEnvironment(env);
-    }
-}
 
 [SetUpFixture]
 public class Integration
 {
-    private static readonly ConcurrentBag<HttpClient> _httpClients = new();
-    public static TestServer TestServer { get; private set; }
-    public static AuthData? AuthData { get; private set; }
+    private static readonly ConcurrentBag<HttpClient> HttpClients = [];
+    public static TestServer TestServer { get; private set; } = null!;
+    public static AuthData AuthData { get; private set; } = null!;
 
     public static HttpClient GetHttpClient()
     {
         HttpClient client = TestServer.CreateClient();
-        _httpClients.Add(client);
+        HttpClients.Add(client);
         return client;
     }
 
-    [OneTimeSetUp]
-    public async Task OneTimeSetUp()
-    {
-        CustomWebApplicationFactory<Program> webHostBuilder = new CustomWebApplicationFactory<Program>();
-        TestServer = webHostBuilder.Server;
-
-       AuthData = await GetAuthData();
-    }
-
-    public static async Task<AuthData?> GetAuthData()
+    public static async Task<AuthData> GetAuthData()
     {
         string username = "bob217@mail.ru";
         string password = "stringA123!";
@@ -67,7 +31,16 @@ public class Integration
         ]);
 
         HttpResponseMessage response = await GetHttpClient().PostAsync($"{TestServer.BaseAddress}connect/token", requestContent);
-        return await response.Content.ReadFromJsonAsync<AuthData>();
+        return await response.Content.ReadFromJsonAsync<AuthData>() ?? throw new InvalidOperationException();
+    }
+
+    [OneTimeSetUp]
+    public async Task OneTimeSetUp()
+    {
+        CustomWebApplicationFactory<Program> webHostBuilder = new();
+        TestServer = webHostBuilder.Server;
+
+        AuthData = await GetAuthData();
     }
 
     [OneTimeTearDown]
@@ -75,17 +48,9 @@ public class Integration
     {
         TestServer.Dispose();
 
-        foreach (HttpClient httpClient in _httpClients.ToArray())
+        foreach (HttpClient httpClient in HttpClients.ToArray())
         {
             httpClient.Dispose();
         }
     }
 }
-
-public record AuthData(
-    [property: JsonPropertyName("access_token")]
-    string AccessToken,
-    [property: JsonPropertyName("token_type")]
-    string TokenType,
-    [property: JsonPropertyName("expires_in")]
-    int ExpiresIn);
